+++ app/api/admin/products/[id]/route.ts
@@ -1,10 +1,11 @@
-import { ProductService } from "@/lib/Services/admin/ProductService";
+﻿import { ProductService } from "@/lib/Services/admin/ProductService";
 import { NextRequest, NextResponse } from "next/server";
+import { AdminAuthrequire } from "@/lib/Services/Auth/AdminAuthrequire";
 
-export async function GET(
-  req: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
-) {
+export async function GET(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
+  const auth = AdminAuthrequire(req);
+  if (!auth.success) return auth.response!;
+
   try {
     const { id } = await params;
     const product = await ProductService.getProductById(id);
@@ -14,10 +15,10 @@ export async function GET(
   }
 }
 
-export async function PUT(
-  req: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
-) {
+export async function PUT(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
+  const auth = AdminAuthrequire(req);
+  if (!auth.success) return auth.response!;
+
   try {
     const { id } = await params;
     const formData = await req.formData();
@@ -28,10 +29,10 @@ export async function PUT(
   }
 }
 
-export async function DELETE(
-  req: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
-) {
+export async function DELETE(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
+  const auth = AdminAuthrequire(req);
+  if (!auth.success) return auth.response!;
+
   try {
     const { id } = await params;
     await ProductService.deleteProduct(id);
